Consolidating conftests and util functionality
